Update Page Sequence of Viewer to Make Pages Work Correctly
